requirements.txt
web3
flask
python-dotenv
requests
kafka-python
google-cloud-bigquery
numpy
praw
pandas
apache-airflow